Prevent monsters from respawning inside each other (sf bug #1793597)
Boom changed P_CheckPosition to allow non-solid things to move through
solid ones. Therefore a respawn of a corpse (non-solid) will not be
prevented, even if its start position is occupied by a solid thing such
as another monster. At this point the corpse respawns, becomes solid
itself, and the two monsters are stuck together.

The easiest fix, which is also used in Eternity, is to pretend the corpse is
solid while calling P_CheckPosition in P_NightmareRespawn.