m_misc.[ch]: macro version of M_DoubleToInt
On GCC, use an extension (compound statement expressions[1]) to declare
a temporary variable; on others, just cast the parameter to double.

I'm not sure why casting the return value of a function from double to
int causes a warning, but putting the result in a temporary double and
casting that to int, does not.

[1] http://gcc.gnu.org/onlinedocs/gcc/Statement-Exprs.html